+++ conformance/interfaces/mmap/6-5.c
@@ -40,7 +40,7 @@ int main(void)
 	int fd;
 
 	/* Create the file */
-	snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_6_5_%d", getpid());
+	snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_6_5_%ld", (long)getpid());
 	unlink(tmpfname);
 	fd = open(tmpfname, O_CREAT | O_RDWR | O_EXCL, S_IRUSR | S_IWUSR);
 	if (fd == -1) {
