+++ conformance/interfaces/sem_open/1-4.c
@@ -29,7 +29,7 @@ int main(void)
 	sem_t *mysemp;
 	char semname[28];
 
-	sprintf(semname, "/" FUNCTION "_" TEST "_%d", getpid());
+	sprintf(semname, "/" FUNCTION "_" TEST "_%ld", (long)getpid());
 
 	mysemp = sem_open(semname, O_CREAT, 0777, 1);
 	if (mysemp == SEM_FAILED || mysemp == NULL) {
